Consolidate exceptions
As suggested in https://github.com/gratipay/aspen-python/issues/448.
