+++ aspen/exceptions.py
@@ -12,10 +12,21 @@ from __future__ import unicode_literals
 from aspen import Response
 
 
+class ConfigurationError(StandardError):
+    """This is an error in any part of our configuration.
+    """
+
+    def __init__(self, msg):
+        StandardError.__init__(self)
+        self.msg = msg
+
+    def __str__(self):
+        return self.msg
+
+
 class LoadError(Exception):
     """Represent a problem loading a resource.
     """
-    # Define this here to avoid import issues when json doesn't exist.
 
 
 class CRLFInjection(Response):
